Add check task duration logic
Calculates the duration of a tasks if both its start and end time have been included in the log file. If the duration is larger than the warning threshold of 5 minutes, it must be included in the report. The code that does that will follow. In either case, the information about the task is removed from the entries object to save memory.