Add finish report logic
Iterates over the remaining tasks in the entries object. These have either started before the time on the first line of the log or have ended after the time on the last line of log. If their duration compared to the time on the first and last line of the report exceeds the warning or error threshold, the respective status is included in report. For the others a note that it is impossible to tell if the thresholds have beed exceeded is included in the report.